moved class C in to own file, also changed its name
