+++ components/Contacts/ContactsScreen.js
@@ -3,7 +3,7 @@ import {AsyncStorage, Button, StyleSheet, ScrollView, View,} from "react-native"
 import Icon from "react-native-vector-icons/Ionicons";
 import randomColor from "randomcolor";
 import ContactList from './ContactList';
-
+import ContactPersistence from './ContactPersistence.js/';
 
 
 /**
@@ -39,8 +39,8 @@ export default class ContactsScreen extends Component {
    */
   fetchContacts = async () => {
     try {
-      const contactIds = await C.allIds();
-      const contacts = await C.getContactsByIds(contactIds);
+      const contactIds = await ContactPersistence.allIds();
+      const contacts = await ContactPersistence.getContactsByIds(contactIds);
       this.setState({contacts});
     } catch (e) {
       console.error(e);
@@ -70,7 +70,7 @@ export default class ContactsScreen extends Component {
         return {contacts};
       });
       // save contact to persistent storage
-      await C.saveContact(myContact);
+      await ContactPersistence.saveContact(myContact);
     } catch (e) {
       console.error(e);
     }
@@ -88,7 +88,7 @@ export default class ContactsScreen extends Component {
       return { contacts: contacts };
       }
     );
-    C.deleteContact(c);
+    ContactPersistence.deleteContact(c);
   };
 
   /**
@@ -131,88 +131,6 @@ export default class ContactsScreen extends Component {
   }
 }
 
-/**
- * This class has static methods for handling persistent storage
- *
- * The classname is terse to make code more readable
- */
-class C {
-
-  /**
-   * Save a contact object to persistent storage
-   *
-   * The contact object must have a key property
-   *
-   * @param contact the contact to save
-   * @returns {Promise<void>} the returned promise can be ignored
-   */
-
-  static async saveContact(contact) {
-    const ids = await C.allIds();
-    if (!ids.includes(contact.key)) {
-      ids.push(contact.key);
-      await AsyncStorage.setItem("CONTACT_IDS", JSON.stringify(ids));
-    }
-    await AsyncStorage.setItem("contact_" + contact.key, JSON.stringify(contact));
-  }
-
-  /**
-   * Get a list of previously saved contact ids from persistent storage
-   * @returns {Promise<*>} A promise resolving to an array of ids
-   */
-  static async allIds() {
-    const idsJson = await AsyncStorage.getItem("CONTACT_IDS");
-    if (!idsJson) {
-      const initialIdsList = [];
-      await AsyncStorage.setItem("CONTACT_IDS", JSON.stringify(initialIdsList));
-      return initialIdsList;
-    }
-    return JSON.parse(idsJson);
-  }
-
-  /**
-   * Get multiple contacts from persistent storage, based on the ids list
-   *
-   * @param ids a list of ids to fetch from persistent storage
-   * @returns {Promise<any[]>} a promise resolving to an array of contact objects
-   */
-  static async getContactsByIds(ids) {
-    const keys = ids.map(id => "contact_" + id);
-    const contactsJson = [];
-    const contactsKeyValue = await AsyncStorage.multiGet(keys);
-    contactsKeyValue
-      .map(element => element[1])
-      .forEach(element => contactsJson.push(element));
-    return contactsJson.map(elementJson => JSON.parse(elementJson));
-  }
-
-
-  static deleteContact = async (c) => {
-    try {
-      //get all IDS
-      const ids = await AsyncStorage.getItem("CONTACT_IDS") || "[]";
-
-      //make them to objects
-      const keys = JSON.parse(ids);
-
-      //Find index of contact we want to delete
-      let index;
-      if ((index = keys.findIndex(el => el === c.key)) !== -1) {
-        //delete it
-        keys.splice(index, 1);
-      }
-
-
-      // Set new contact_ids
-      await AsyncStorage.setItem("CONTACT_IDS", JSON.stringify(keys));
-
-      //remove key from asyncstorage
-      await AsyncStorage.removeItem(c.key.toString());
-    } catch (e) {
-      console.error(e);
-    }
-  };
-}
 
 const styles = StyleSheet.create({
   container: {
